requirements.txt
Flask==1.1.1
numpy==1.19.4
matplotlib==3.2.2
torch==1.8.2
torchvision==0.10.1
Pillow==7.1.0
flask_cors==3.0.10
opencv-python==4.5.1.48
wandb==0.10.12
tqdm==4.55.1
helper==2.4.2
scipy==1.7.1
facenet-pytorch==2.5.2
